# docs/swagger.yaml
basePath: /api/v1
definitions:
  http.AuthResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  http.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  http.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  http.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  http.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      timezone:
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/davidfolch/apocapoc-api
  description: Self-hosted habit tracking service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Apocapoc API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password. Returns both access
        token and refresh token. The access token is used for API requests, the refresh
        token is used to obtain new access tokens when they expire.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns access token, refresh token, and user ID
          schema:
            $ref: '#/definitions/http.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the refresh token to invalidate the user session. After
        logout, the refresh token cannot be used to obtain new access tokens. The
        user will need to login again. Always call this endpoint before clearing tokens
        from client storage to ensure proper session termination.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out - the refresh token is now invalid
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or missing refresh token
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Refresh token not found (already revoked or never existed)
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 'Exchange a valid refresh token for a new access token and refresh
        token pair. IMPORTANT: The old refresh token is automatically revoked and
        you receive a NEW refresh token - always update both tokens in storage. Use
        this endpoint when the access token expires to maintain the user session without
        requiring re-login.'
      parameters:
      - description: Current refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns NEW access token and NEW refresh token - the old refresh
            token is now invalid
          schema:
            $ref: '#/definitions/http.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and receive both access token and refresh
        token. Store both tokens securely - the refresh token is used to obtain new
        access tokens when they expire.
      parameters:
      - description: 'Registration data (password requires: min 8 chars, uppercase,
          lowercase, digit, special char)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Returns access token, refresh token, and user ID
          schema:
            $ref: '#/definitions/http.AuthResponse'
        "400":
          description: 'Invalid input: email format, password requirements, or timezone'
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"